feat: migrate matching script logic to actions
