+++ agir/voting_proxies/actions.py
@@ -1,10 +1,14 @@
 from copy import deepcopy
+from datetime import timedelta
 from functools import partial
 
+from data_france.models import Commune
 from django.contrib.gis.db.models.functions import Distance
+from django.contrib.gis.measure import D
 from django.contrib.postgres.aggregates import ArrayAgg
 from django.db import transaction
-from django.db.models import Count
+from django.db.models import Count, Q
+from django.utils import timezone
 
 from agir.lib.tasks import geocode_person
 from agir.people.actions.subscription import (
@@ -13,12 +17,18 @@ from agir.people.actions.subscription import (
     SUBSCRIPTION_TYPE_AP,
 )
 from agir.people.models import Person
+from agir.voting_proxies import tasks
 from agir.voting_proxies.models import VotingProxy, VotingProxyRequest
 from agir.voting_proxies.tasks import (
     send_voting_proxy_request_confirmation,
     send_voting_proxy_request_accepted_text_messages,
     send_voting_proxy_request_confirmed_text_messages,
 )
+
+
+# TODO: Choose a proxy-to-request distance limit (in meters)
+PROXY_TO_REQUEST_DISTANCE_LIMIT = 30000  # 30 KM
+PER_VOTING_PROXY_REQUEST_INVITATION_LIMIT = 10
 
 
 def create_or_update_voting_proxy_request(data):
@@ -121,10 +131,6 @@ def update_voting_proxy(instance, data):
     return instance
 
 
-# TODO: Choose a proxy-to-request distance limit (in meters)
-PROXY_TO_REQUEST_DISTANCE_LIMIT = 30000  # 30 KM
-
-
 def get_voting_proxy_requests_for_proxy(voting_proxy, voting_proxy_request_pks):
     voting_proxy_requests = VotingProxyRequest.objects.filter(
         status=VotingProxyRequest.STATUS_CREATED,
@@ -205,3 +211,152 @@ def confirm_voting_proxy_requests(voting_proxy_requests):
     voting_proxy_request_pks = list(voting_proxy_requests.values_list("pk", flat=True))
     voting_proxy_requests.update(status=VotingProxyRequest.STATUS_CONFIRMED)
     send_voting_proxy_request_confirmed_text_messages.delay(voting_proxy_request_pks)
+
+
+def send_matching_requests_to_proxy(proxy, matching_request_ids):
+    proxy.last_matched = timezone.now()
+    proxy.save()
+    tasks.send_matching_request_to_voting_proxy.delay(
+        proxy.id, list(matching_request_ids)
+    )
+
+
+def match_available_proxies_with_requests(
+    pending_requests, notify_proxy=send_matching_requests_to_proxy
+):
+    fulfilled_request_ids = []
+
+    # Retrieve all available proxy that has not been matched in the last two days
+    available_proxies = VotingProxy.objects.filter(
+        status__in=(VotingProxy.STATUS_CREATED, VotingProxy.STATUS_AVAILABLE),
+    ).exclude(last_matched__date__gt=timezone.now() - timedelta(days=2))
+
+    # Try to match available voting proxies with pending requests
+    for proxy in available_proxies:
+        pending_requests = pending_requests.exclude(id__in=fulfilled_request_ids)
+        if not pending_requests.exists():
+            break
+        try:
+            matching_request_ids = get_voting_proxy_requests_for_proxy(
+                proxy, pending_requests.values_list("id", flat=True)
+            ).values_list("id", flat=True)
+        except VotingProxyRequest.DoesNotExist:
+            pass
+        else:
+            notify_proxy(proxy, matching_request_ids)
+            fulfilled_request_ids += matching_request_ids
+
+    return fulfilled_request_ids
+
+
+def invite_voting_proxy_candidates(candidates):
+    voting_proxy_candidates = [
+        VotingProxy(
+            status=VotingProxy.STATUS_INVITED,
+            person=candidate,
+            email=candidate.email,
+            first_name=candidate.first_name if candidate.first_name else "-",
+            last_name=candidate.last_name if candidate.first_name else "-",
+            contact_phone=candidate.contact_phone if candidate.first_name else "-",
+            polling_station_number="",
+        )
+        for candidate in candidates
+    ]
+    voting_proxy_candidates = VotingProxy.objects.bulk_create(
+        voting_proxy_candidates, ignore_conflicts=True
+    )
+    voting_proxy_candidates_ids = [
+        voting_proxy_candidate.pk for voting_proxy_candidate in voting_proxy_candidates
+    ]
+    tasks.send_voting_proxy_candidate_invitation_email.delay(
+        voting_proxy_candidates_ids
+    )
+
+    return voting_proxy_candidates_ids
+
+
+def get_voting_proxy_candidates_queryset(request, blacklist_ids):
+    candidates = (
+        Person.objects.exclude(
+            id__in=VotingProxy.objects.values_list("person_id", flat=True)
+        )
+        .exclude(emails__address=None)
+        .filter(is_2022=True, newsletters__len__gt=0)
+    )
+    if request and request["email"]:
+        candidates = candidates.exclude(emails__address=request["email"])
+    if blacklist_ids:
+        candidates = candidates.exclude(id__in=blacklist_ids)
+
+    return candidates
+
+
+def find_voting_proxy_candidates_for_requests(
+    pending_requests, send_invitations=invite_voting_proxy_candidates
+):
+    possibly_fulfilled_request_ids = []
+    candidate_ids = []
+
+    # Find candidates for consulate requests
+    for request in (
+        pending_requests.exclude(consulate__pays__isnull=True)
+        .values("email", "consulate__pays")
+        .annotate(ids=ArrayAgg("id"))
+    ):
+        candidates = get_voting_proxy_candidates_queryset(
+            request, candidate_ids
+        ).filter(location_country__in=request["consulate__pays"].split(","))[
+            :PER_VOTING_PROXY_REQUEST_INVITATION_LIMIT
+        ]
+
+        if candidates.exists():
+            invited_candidate_ids = send_invitations(candidates)
+            candidate_ids += invited_candidate_ids
+            possibly_fulfilled_request_ids += request["ids"]
+
+    # Find candidates for commune requests
+    for request in (
+        pending_requests.exclude(commune__isnull=True)
+        .values(
+            "email",
+            "commune__id",
+            "commune__mairie_localisation",
+        )
+        .annotate(ids=ArrayAgg("id"))
+    ):
+        candidates = get_voting_proxy_candidates_queryset(
+            request, candidate_ids
+        ).exclude(coordinates__isnull=True)
+
+        # Match by distance for geolocalised communes
+        if request["commune__mairie_localisation"]:
+            candidates = candidates.filter(
+                coordinates__dwithin=(
+                    request["commune__mairie_localisation"],
+                    D(m=PROXY_TO_REQUEST_DISTANCE_LIMIT),
+                )
+            )
+        # Try to match by city code / zip code for non-geolocalised communes
+        else:
+            commune = Commune.objects.get(id=request["commune__id"])
+            candidates = candidates.filter(
+                Q(location_citycode=commune.code) | Q(zip__in=commune.codes_postaux)
+            ).distinct()
+
+        if candidates.exists():
+            # Give priority with people with the most recent event rsvps
+            candidates = candidates.annotate(
+                rsvp_count=Count(
+                    "rsvps",
+                    filter=Q(
+                        rsvps__event__end_time__gte=timezone.now() - timedelta(days=365)
+                    ),
+                    distinct=True,
+                )
+            ).order_by("-rsvp_count")[:PER_VOTING_PROXY_REQUEST_INVITATION_LIMIT]
+
+            invited_candidate_ids = send_invitations(candidates)
+            candidate_ids += invited_candidate_ids
+            possibly_fulfilled_request_ids += request["ids"]
+
+    return possibly_fulfilled_request_ids, candidate_ids
